+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - A121 presence: Fix depth filter length so it never is bigger than
   the number of points
+- Correct downloading of new firmware by adapting to new developer site
 
 ## v5.5.1